feat : 다른 페이지 이동 시, mapCenter 기억해둠
